Bug fix for scheduler, new variables in packet
1) The Python sched object will sometimes not schedule events in the
order in which they were added to the heap, if the events have the
same time. I added a counter for the priority field of an event, so events
scheduled at the same time are now handled in the order in which they
were added to the system. This primarily affects adding a large number
of packets at the start of the simulation.

2) Added the ability to track total transmission delay and propagation
delay for each packet.