Add initial Python 3 support
Most changes were made with 2to3.
The majority of changes were to the print function and division.
It doesn't really seem to depend on binary data much?
More revision may be needed, and unit tests created for verification.

The `from __future__ import print_function` is added to keep Python 2
compatibility for now considering it was very easy to do.

Resolves #6.